requirements.txt
altair==5.4.1
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.1.2
bidict==0.23.1
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==43.0.1
cssbeautifier==1.15.1
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
djlint==1.34.1
EditorConfig==0.12.4
executing==2.1.0
filelock==3.16.1
firebase-admin==6.5.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.2
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.25.1
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.3
jiter==0.5.0
jsbeautifier==1.15.1
json5==0.9.24
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keyboard==0.13.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.9.1
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.9.0
PyMsgBox==1.0.9
pyparsing==3.1.4
PyPDF2==3.0.1
pyperclip==1.8.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pytweening==1.0.7
pytz==2024.2
pywin32==306
PyYAML==6.0.1
pyzmq==26.2.0
qasync==0.27.1
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
SecretStorage==3.3.3
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SpeechRecognition==3.10.4
SQLAlchemy==2.0.29
stack-data==0.6.3
streamlit==1.39.0
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.20.0
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
transformers==4.45.1
typing_extensions==4.10.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==5.0.3
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.1
wsproto==1.2.0